% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-plots.R
\name{calculate_performance}
\alias{calculate_performance}
\title{Calculate classifier performance}
\usage{
calculate_performance(score, labels, plot_index, plot.colours,
  add_to_plot = FALSE, do_not_plot = FALSE)
}
\arguments{
\item{score}{vector of scores, one for each model (e.g. RFQAmodel)}

\item{labels}{vector of labels, one for each model (1 for correct, 0 for
incorrect)}
}
\description{
Calculates the pr, auc and performance for a given score on a set of models
}
