% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-files.R
\name{collect_input}
\alias{collect_input}
\title{Collect input files}
\usage{
collect_input(directory = ".", detailsfile = "target_details.txt",
  targetnamelength = 6)
}
\arguments{
\item{directory}{string. The directory containing files (default: working)}

\item{detailsfile}{string. The file containing details about the targets
(default: "target_details.txt")}
}
\value{
Returns a dataframe containing the raw data for the targets.
}
\description{
Collect the raw feature information from files
}
