% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-files.R
\name{gather_files}
\alias{gather_files}
\title{Gather files}
\usage{
gather_files(extension, colnames, directory = ".")
}
\arguments{
\item{extension}{string. File extension}

\item{colnames}{vector. Column names}
}
\value{
Returns a data frame containing data from the specified files for
all the targets in the current directory
}
\description{
Collect the features produced by the get_features.sh script
}
