% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-results.R
\name{get_confidence}
\alias{get_confidence}
\title{Get confidence categories}
\usage{
get_confidence(results, predictor = RFQAmodel,
  confidence_cutoffs = c(0.5, 0.3, 0.1))
}
\arguments{
\item{predictor}{string. Name of the column to use as the score}

\item{confidence_cutoffs}{vector. Optional custom confidence cutoffs}
}
\description{
Classify targets into confidence categories according to the RFQAmodel score
of the highest-ranking model
}
