% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-files.R
\name{get_details}
\alias{get_details}
\title{Get target details}
\usage{
get_details()
}
\arguments{
\item{string.}{Name of the file containing target-specific details}
}
\value{
Returns a dataframe with the target details
}
\description{
Read in the target details from a specified file
}
\details{
The target details file should include the following columns:
Target, Beff, Length

The header should be included.
}
