% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-features.R
\name{get_ensemble}
\alias{get_ensemble}
\title{Get ensemble features}
\usage{
get_ensemble(df, colname, prefix = colname, rev = FALSE)
}
\arguments{
\item{df}{dataframe. The features dataset to calculate from and append to}

\item{colname}{string. The name of the column corresponding to the feature
for which ensemble features will be calculated}

\item{prefix}{string. The prefix used when naming the ensemble feature
columns (defaults to the name of the column)}

\item{rev}{boolean. Flag to indicate that smaller scores are better
for this feature}
}
\value{
Returns a dataframe that includes the new ensemble features
}
\description{
Calculate the max, minimum, median and spread for a given score
}
\examples{
get_ensemble(features, "ProQ2D") \%>\%
get_ensemble(., "SAINT2", "S2", rev=TRUE)
}
