% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-features.R
\name{get_features}
\alias{get_features}
\title{Get the features}
\usage{
get_features(df, scaffold = FALSE, noglobal = FALSE)
}
\arguments{
\item{df}{dataframe. Data frame containing the raw input data}

\item{scaffold}{boolean. Indicate whether to use ScafFold features.
Defaults to FALSE, use standard RFQAmodel features}

\item{noglobal}{boolean. Indicate whether to leave out global features}
}
\value{
Returns a dataframe containing the features for RFQAmodel
}
\description{
Get the features from the input data
}
