% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-features.R
\name{get_labels}
\alias{get_labels}
\title{Get labels}
\usage{
get_labels(truth = TMScore, cutoff = 0.5, rev = FALSE)
}
\arguments{
\item{truth}{vector. Model quality compared to native (e.g. TMScore) for all
models}

\item{cutoff}{float. The cutoff that defines a correct model (default 0.5)}
}
\value{
Returns a vector of the length of the number of models, with 1 or 0
indicating correct or incorrect
}
\description{
Label individual models as correct or incorrect
}
