% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-results.R
\name{get_stats}
\alias{get_stats}
\title{Get results statistics}
\usage{
get_stats(results, predictor = RFQAmodel, relabel = FALSE,
  truth = "TMScore", cutoff = 0.5, rev = FALSE)
}
\arguments{
\item{results}{dataframe}

\item{predictor}{string. Column to use as the predictor (default RFQAmodel)}

\item{truth}{string. Column to use as the truth (default TMScore)}

\item{cutoff}{float. Cutoff that defines a correct model (default 0.5)}

\item{rev}{boolean. True if lower values are better for the truth (e.g. RMSD) (default FALSE)}
}
\description{
Get a table of results for how well RFQAmodel performed for a set
}
