% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-plots.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot a ROC curve}
\usage{
plot_roc(resultsdf, featurelist, topmodel = FALSE, addhighconf = FALSE,
  truth = TMScore, cutoff = 0.5)
}
\arguments{
\item{resultsdf}{data frame}

\item{featurelist}{vector. List of methods/features to plot curves for}

\item{topmodel}{boolean. When true, only the highest-ranking model per
target is considered}

\item{addhighconf}{boolean. When true, the TPR and FPR corresponding to the
high confidence cutoff (0.5) is indicated}

\item{truth}{string. Optional, the score to use to define a correct model
(default TMScore). This is only required if the Label column is not already
provided.}

\item{cutoff}{float. Optional, the cutoff to use to define a correct models.
See \code{truth}.}
}
\description{
Plots a ROC curve for one or more methods
}
